intellikat/pyproject.toml
[project]
name = "intellikat"
version = "0.1.0"
description = "Rekall Intellikat — sentiment + summarization microservice for transcripts"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "Proprietary" }
authors = [{ name = "Rekall" }]

dependencies = [
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.1",
    "alembic>=1.13",
    "azure-servicebus>=7.12",
    "azure-identity>=1.16",
    "httpx>=0.27",
    "openai>=1.40",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-json-logger>=2.0",
]

[project.optional-dependencies]
local = [
    "transformers>=4.40",
    "torch>=2.2",
    "tokenizers>=0.19",
    "accelerate>=0.30",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "testcontainers[postgres]>=4.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "import-linter>=2.0",
    "bandit>=1.7",
]

[project.scripts]
intellikat = "intellikat.main:main"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/intellikat"]

[tool.ruff]
line-length = 110
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "PL", "RUF"]
ignore = ["PLR0913", "PLR2004"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["PLR2004", "PLR0915"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
mypy_path = "src"
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["transformers.*", "torch.*", "azure.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
pythonpath = ["src"]
markers = [
    "integration: tests that require Postgres / Service Bus / network",
    "local_hf: tests that require the [local] extras and download model weights",
]
